feat: PageEvents defined
