+++ apps/page-service/src/page-service.module.ts
@@ -2,11 +2,12 @@ import { Module } from '@nestjs/common';
 import { CategoryService } from './category.service';
 import { PageServiceController } from './page-service.controller';
 import { PageService } from './page-service.service';
+import { PostService } from './post-service.service';
 
 @Module({
   imports: [],
   controllers: [PageServiceController],
-  providers: [PageService, CategoryService],
-  exports: [PageService, CategoryService],
+  providers: [PostService, CategoryService, PageService],
+  exports: [PostService, CategoryService, PageService],
 })
 export class PageServiceModule {}
